Some changes to get webpack to not include all calendar types.
It was processing a require() that was only used for dynamic
code loading and which allowed for modules that weren't there.
Webpack was including everything just in case it was ever
needed, which is not what we want. The Calendars should be
like plugins. If they are there, we can use them. If not, then
it should still work too. Gregorian is the only calendar that
must be there.