Emergency patch: ilib 13.1.0 in npm was broken!
The problem was that the uglify run quietly failed and did not
stop the build. This means that the compressed js files did not
get included into the tar file that was published to npm as 13.1.0.
This fix temporarily fixes that, and we can publish again as 13.1.1
with no code changes. Only build fixes. All the build stuff is
reworked anyways in the build.xml in the webpack branch, so we
will not need this patch in development.