Fix async html tests.
We can't do real async tests with the asembled ilib yet
because the dynamic loading of data depends on webpack.
Instead, here we use an assembled version of ilib and ensure
that the calls with async callbacks work correctly. These
tests will become more useful when the webpack support is
added.