Dummy source file and project file to get localizations
Using the loctool, we can extract all the strings into xliff format.
The xliff format was then localized by professionals, and the result
went into ilib.xliff. This is then used to generate the localized
js files which I changed into json files by hand and checked in to
the locale directory.