Work around problem where uglify was removing the !defineLocaleData comment
Without the !defineLocaleData comment, the plugin and loader do not have
any target where they could put their output and therefore there is no
dependency on the locale data, which therefore never gets included into
the webpacked code.

The problem was that webpack thought the code block with only the comment
in it was empty, so the dead-code optimizer removed that entire block,
including the comment. The solution was to put the comment in another
file and require that file in that block so that the dead code optimizer
does not think that block is empty.