More work on getFormatInfo
Add the idea of a value property, the value of which is a
function that calculates the value of this field based on
the value of other fields. For example, the day of week is
not something that a UI typically displays in a date input
form. Instead, you pick the year, month, and day and the
day of week is determined automatically.

More work is needed for time zones and to translate stuff.