Work around problem in cldr-data
The json was generated with an extra space in the currency format
for fr-CH. It seems to be a conversion error between the CLDR
XML, which is canonical, and the json in cldr-data, which is
derivative. The XML did not have the space, and the json data
introduced it.

The solution was to trim the white space from the
start and end of all number formats to work around that problem.