Resolve the file name cache fix differently
If you don't check the fileSet cache, then it will
attempt to load the same missing json files every
time. That is, if a json file does not exist for a locale,
then attempting to load it will fail, and nothing
stops it from trying again and again to load that same
missing file every time it is requested.