Add in the missing methods
These methods have been added to the String prototype
in the years since this class was first coded. Here we
put in methods to delegate to the string class. The unit
tests only attempt to run if the string class supports
these. Older versions of node and browsers of course do
not have these new methods.