Fixed an edge case where a format is missing
In the case where the zone's format (zone abbreviation)
is simply "%s" and the zone has changed from having DST
in the past but not now, then there is no current rule with
which to replace the "%s". The work-around is to find the
last rule that is available and use its format letters as
the format.