Fix the day-of-the-week calculations for Ethiopic
The calculations were set to calculate the correct day of
the week for Gregorian midnight. But, Ethiopic dates start at
6am of a Gregorian day, so the calculation for midnight should
be one day less. Take the Gregorian day Aug 20, 2021 as an
example. That day is a Friday. In the Ethiopic calendar, from
midnight to 6am of that Gregorian day, it is Thursday,
13 Nehasse 2013. The date and day-of-the-week switch at 6am
Gregorian time to Friday, 14 Nehasse 2013.